Migración a Postgress y Ajustes de Carga

+++ model/gestionar_db.py
@@ -1,11 +1,11 @@
-import sqlite3
-from sqlite3 import OperationalError
+import psycopg2
+from psycopg2 import OperationalError
 from threading import Lock
 from datetime import datetime, time
 import pytz
 from fastapi import HTTPException
 
-DATABASE_PATH = "./centro_control.db"
+DATABASE_PATH = "postgresql://gestionexpress:G3st10n3xpr3ss@serverdbcexp.postgres.database.azure.com:5432/gestionexpress" 
 # Establecer la zona horaria de Colombia
 colombia_tz = pytz.timezone('America/Bogota')
 
@@ -17,7 +17,7 @@ class HandleDB:
         with cls._lock:
             if not cls._instance:
                 cls._instance = super().__new__(cls)
-                cls._instance._con = sqlite3.connect("./centro_control.db", check_same_thread=False)  # Habilita multi-hilo
+                cls._instance._con = psycopg2.connect(DATABASE_PATH)
                 cls._instance._cur = cls._instance._con.cursor()
             return cls._instance
 
@@ -26,15 +26,22 @@ class HandleDB:
             self._cur.execute("SELECT * FROM usuarios")
             return self._cur.fetchall()
 
-    def get_only(self, username): # Método para obtener un único registro de usuario filtrado por nombre de usuario
+    def get_only(self, username): 
         with self._lock:
-            self._cur.execute("SELECT * FROM usuarios WHERE username = ?", (username,))
-            return self._cur.fetchone()
-
+            cur = self._con.cursor()  # Crear un nuevo cursor
+            try:
+                cur.execute("SELECT * FROM usuarios WHERE username = %s", (username,))
+                result = cur.fetchone()
+            finally:
+                cur.close()  # Asegurarse de cerrar el cursor después de su uso
+            return result
+                
     def insert(self, data_user): # Método para insertar un nuevo usuario en la base de datos
         with self._lock:
-            self._cur.execute("INSERT INTO usuarios VALUES (?, ?, ?, ?, ?, ?)", (
-                data_user["id"],
+            self._cur.execute("""
+                INSERT INTO usuarios (nombres, apellidos, username, cargo, password_user)
+                VALUES (%s, %s, %s, %s, %s)
+            """, (
                 data_user["nombres"],
                 data_user["apellidos"],
                 data_user["username"],
@@ -49,24 +56,52 @@ class HandleDB:
 
 class Cargue_Controles:
     def __init__(self):
-        self.database_path = "./centro_control.db"
-        self.conn = sqlite3.connect(self.database_path, timeout=10)
+        self.database_path = "postgresql://gestionexpress:G3st10n3xpr3ss@serverdbcexp.postgres.database.azure.com:5432/gestionexpress"
+        self.conn = psycopg2.connect(self.database_path)
         self.cursor = self.conn.cursor()
-
-    def borrar_tablas(self):
-        self.cursor.execute('DELETE FROM planta')
-        self.cursor.execute('DELETE FROM controles')
-        self.cursor.execute('DELETE FROM supervisores')
-        self.cursor.execute('DELETE FROM turnos')
-        self.conn.commit()
+        
+    def borrar_tablas(self, tablas_a_borrar):
+        try:
+            if 'planta' in tablas_a_borrar:
+                self.cursor.execute('DELETE FROM planta')
+            if 'supervisores' in tablas_a_borrar:
+                self.cursor.execute('DELETE FROM supervisores')
+            if 'turnos' in tablas_a_borrar:
+                self.cursor.execute('DELETE FROM turnos')
+            if 'controles' in tablas_a_borrar:
+                self.cursor.execute('DELETE FROM controles')
+            self.conn.commit()
+        except Exception as e:
+            self.conn.rollback()
+            print(f"Error al borrar las tablas: {str(e)}")
+            raise
 
     def cargar_datos(self, data):
-        self.borrar_tablas()  # Borrar los datos existentes antes de cargar los nuevos
+        # Crear una lista de las tablas a borrar según las hojas seleccionadas
+        tablas_a_borrar = []
+        if 'planta' in data:
+            tablas_a_borrar.append('planta')
+        if 'supervisores' in data:
+            tablas_a_borrar.append('supervisores')
+        if 'turnos' in data:
+            tablas_a_borrar.append('turnos')
+        if 'controles' in data:
+            tablas_a_borrar.append('controles')
+
+        # Borrar solo las tablas seleccionadas
+        self.borrar_tablas(tablas_a_borrar)
+
         print("Iniciando la carga de datos...")
-        self._cargar_planta(data['planta'])
-        self._cargar_supervisores(data['supervisores'])
-        self._cargar_turnos(data['turnos'])
-        self._cargar_controles(data['controles'])
+
+        if 'planta' in data:
+            self._cargar_planta(data['planta'])
+        if 'supervisores' in data:
+            self._cargar_supervisores(data['supervisores'])
+        if 'turnos' in data:
+            self._cargar_turnos(data['turnos'])
+        if 'controles' in data:
+            self._cargar_controles(data['controles'])
+
         print("Datos cargados exitosamente.")
         self.conn.close()
 
@@ -75,7 +110,7 @@ class Cargue_Controles:
             for row in planta_data:
                 print(f"Insertando o actualizando en planta: {row}")
                 self.cursor.execute('''
-                    INSERT INTO planta (cedula, nombre) VALUES (?, ?)
+                    INSERT INTO planta (cedula, nombre) VALUES (%s, %s)
                     ON CONFLICT(cedula) DO UPDATE SET nombre=excluded.nombre
                 ''', (row['cedula'], row['nombre']))
             self.conn.commit()
@@ -89,7 +124,7 @@ class Cargue_Controles:
             for row in supervisores_data:
                 print(f"Insertando o actualizando en supervisores: {row}")
                 self.cursor.execute('''
-                    INSERT INTO supervisores (cedula, nombre) VALUES (?, ?)
+                    INSERT INTO supervisores (cedula, nombre) VALUES (%s, %s)
                     ON CONFLICT(cedula) DO UPDATE SET nombre=excluded.nombre
                 ''', (row['cedula'], row['nombre']))
             self.conn.commit()
@@ -103,10 +138,10 @@ class Cargue_Controles:
             for row in turnos_data:
                 print(f"Insertando o actualizando en turnos: {row}")
                 self.cursor.execute('''
-                    INSERT OR REPLACE INTO turnos (turno, hora_inicio, hora_fin, detalles)
-                    VALUES (?, ?, ?, ?)
-                    ''', (str(row['turno']), str(row['hora_inicio']), str(row['hora_fin']), str(row['detalles']))
-                )
+                    INSERT INTO turnos (turno, hora_inicio, hora_fin, detalles)
+                    VALUES (%s, %s, %s, %s)
+                    ON CONFLICT(turno) DO UPDATE SET hora_inicio = excluded.hora_inicio, hora_fin = excluded.hora_fin, detalles = excluded.detalles
+                ''', (str(row['turno']), str(row['hora_inicio']), str(row['hora_fin']), str(row['detalles'])))
             self.conn.commit()
         except Exception as e:
             self.conn.rollback()
@@ -114,7 +149,7 @@ class Cargue_Controles:
             raise
 
     def _cargar_controles(self, controles_data):
-        batch_size = 10  # Tamaño del lote
+        batch_size = 450  # Tamaño del lote
         retries = 10  # Número de reintentos
         delay = 1  # Retraso entre reintentos en segundos
 
@@ -122,31 +157,38 @@ class Cargue_Controles:
             batch = controles_data[i:i + batch_size]
             for _ in range(retries):
                 try:
-                    conn = sqlite3.connect(self.database_path, timeout=10)
+                    conn = psycopg2.connect(self.database_path)
                     cursor = conn.cursor()
-                    cursor.execute('PRAGMA busy_timeout = 30000')  # Establecer tiempo de espera de 30 segundos
 
                     for row in batch:
                         print(f"Insertando o actualizando en controles: {row}")
                         cursor.execute('''
-                            INSERT OR REPLACE INTO controles (concesion, puestos, control, ruta, linea, admin, cop, tablas) 
-                            VALUES (?, ?, ?, ?, ?, ?, ?, ?)
+                            INSERT INTO controles (concesion, puestos, control, ruta, linea, admin, cop, tablas) 
+                            VALUES (%s, %s, %s, %s, %s, %s,%s, %s)
+                            ON CONFLICT(id) DO UPDATE SET concesion = excluded.concesion, puestos = excluded.puestos, ruta = excluded.ruta, linea = excluded.linea, admin = excluded.admin, cop = excluded.cop, tablas = excluded.tablas
                         ''', (row['concesion'], row['puestos'], row['control'], row['ruta'], row['linea'], row['admin'], row['cop'], row['tablas']))
-                    conn.commit()
-                    conn.close()
-                    break
-                except sqlite3.OperationalError as e:
-                    if "database is locked" in str(e):
-                        print("La base de datos está bloqueada, reintentando...")
-                        time.sleep(delay)  # Esperar antes de reintentar
-                    else:
-                        raise
+                    
+                    conn.commit()  # Hacer commit si todo va bien
+                    break  # Salir del ciclo de reintentos si todo es exitoso
+
+                except psycopg2.OperationalError as e:
+                    # Manejo de errores específicos
+                    print(f"Error de conexión: {str(e)}, reintentando...")
+                    time.sleep(delay)  # Esperar antes de reintentar
+
+                except Exception as e:
+                    # Manejo de errores generales
+                    conn.rollback()  # Hacer rollback en caso de error
+                    print(f"Error al cargar los datos de controles: {str(e)}")
+                    raise
+
                 finally:
                     if conn:
-                        conn.close()
+                        conn.close()  # Asegurar que la conexión siempre se cierra
+
             else:
-                raise sqlite3.OperationalError("No se desbloquear la base de datos después de varios intentos")
-            
+                raise psycopg2.OperationalError("No se pudo desbloquear la base de datos después de varios intentos")                 
+               
 class Cargue_Asignaciones:
     def __init__(self):
         self.database_path = DATABASE_PATH
@@ -154,18 +196,22 @@ class Cargue_Asignaciones:
     def procesar_asignaciones(self, assignments, user_session):
         processed_data = []
         try:
-            conn = sqlite3.connect(self.database_path, timeout=10)
+            conn = psycopg2.connect(self.database_path, connect_timeout=30)
             cursor = conn.cursor()
+
+            # Recupera todos los controles en una sola consulta y los guarda en un diccionario
+            cursor.execute("SELECT ruta, linea, cop FROM controles")
+            controles = {row[0]: (row[1], row[2]) for row in cursor.fetchall()}
+
             for asignacion in assignments:
                 rutas = asignacion['rutas_asociadas'].split(',')
                 for ruta in rutas:
                     if asignacion['turno'] in ["AUSENCIA", "DESCANSO", "VACACIONES", "OTRAS TAREAS"]:
                         concesion = control = linea = cop = "NO APLICA"
-                        ruta = "NO APLICA"  # No se usa directamente en el dicccionario
+                        ruta = "NO APLICA"
                     else:
-                        cursor.execute("SELECT linea, cop FROM controles WHERE ruta = ?", (ruta.strip(),))
-                        control_data = cursor.fetchone()
-                        linea, cop = control_data if control_data else ("", "")
+                        # Obtén la información del control de la ruta en el diccionario
+                        linea, cop = controles.get(ruta.strip(), ("", ""))
                         concesion = asignacion['concesion']
                         control = asignacion['control']
                         
@@ -178,66 +224,58 @@ class Cargue_Asignaciones:
                         'h_fin': asignacion['hora_fin'],
                         'concesion': concesion,
                         'control': control,
-                        'ruta': ruta if asignacion['turno'] in ["AUSENCIA", "DESCANSO", "VACACIONES", "OTRAS TAREAS"] else ruta.strip(),
+                        'ruta': ruta.strip() if asignacion['turno'] not in ["AUSENCIA", "DESCANSO", "VACACIONES", "OTRAS TAREAS"] else "NO APLICA",
                         'linea': linea,
                         'cop': cop,
                         'observaciones': asignacion['observaciones'],
-                        'usuario_registra': user_session['username'],  # Agregar username
-                        'registrado_por': f"{user_session['nombres']} {user_session['apellidos']}",   # Agregar nombre y apellido
+                        'usuario_registra': user_session['username'],
+                        'registrado_por': f"{user_session['nombres']} {user_session['apellidos']}",
                         'fecha_hora_registro': datetime.now(colombia_tz).strftime("%d-%m-%Y %H:%M:%S"),
-                        'puestosSC': int(asignacion.get('puestosSC', 0)),  # Configuración campo puestosSC
-                        'puestosUQ': int(asignacion.get('puestosUQ', 0)),   # Configuración campo puestosUQ
-                        'cedula_enlace': asignacion['cedula_enlace'],  # Añadir cedula_enlace
-                        'nombre_supervisor_enlace': asignacion['nombre_supervisor_enlace']  # Añadir nombre_supervisor_enlace                        
+                        'puestosSC': int(asignacion.get('puestosSC', 0)),
+                        'puestosUQ': int(asignacion.get('puestosUQ', 0)),
+                        'cedula_enlace': asignacion['cedula_enlace'],
+                        'nombre_supervisor_enlace': asignacion['nombre_supervisor_enlace']
                     })
+
             conn.close()
-        except sqlite3.Error as e:
+        except psycopg2.Error as e:
             print(f"Error al procesar asignaciones: {e}")
         return processed_data
 
     def cargar_asignaciones(self, processed_data):
         try:
-            conn = sqlite3.connect(self.database_path, timeout=10)
+            conn = psycopg2.connect(self.database_path, connect_timeout=30)
             cursor = conn.cursor()
 
-            # Fase de eliminación
-            for data in processed_data:
-                # Asegurarse de que todos los campos obligatorios tienen un valor
-                if not all([data['fecha'], data['cedula'], data['nombre'], data['turno'], data['h_inicio'], data['h_fin']]):
-                    print(f"Datos incompletos para la cédula: {data['cedula']}, data: {data}")
-                    continue
+            # Fase de eliminación en lote (en lugar de hacerlo uno por uno)
+            delete_conditions = [(data['fecha'], data['cedula'], data['turno'], data['h_inicio'], data['h_fin'], data['control'])
+                                 for data in processed_data]
+            cursor.executemany('''
+                DELETE FROM asignaciones
+                WHERE fecha = %s AND cedula = %s AND turno = %s AND h_inicio = %s AND h_fin = %s AND control = %s
+            ''', delete_conditions)
 
-                # Eliminar todos los registros existentes con la misma combinación de (fecha, cédula, turno, h_inicio, h_fin, control)
-                cursor.execute('''
-                    DELETE FROM asignaciones
-                    WHERE fecha = ? AND cedula = ? AND turno = ? AND h_inicio = ? AND h_fin = ? AND control = ?
-                ''', (data['fecha'], data['cedula'], data['turno'], data['h_inicio'], data['h_fin'], data['control']))
-                print(f"Registros eliminados para cédula: {data['cedula']}, fecha: {data['fecha']}, turno: {data['turno']}, h_inicio: {data['h_inicio']}, h_fin: {data['h_fin']}, control: {data['control']}")
+            print(f"Registros eliminados para las asignaciones.")
 
-            # Fase de inserción
-            for data in processed_data:
-                # Desglosar las rutas en filas separadas e insertar los nuevos registros
-                rutas = data['ruta'].split(',')
-                for ruta in rutas:
-                    ruta = ruta.strip()  # Asegurarse de que no hay espacios innecesarios
+            # Fase de inserción en lote
+            insert_values = [
+                (data['fecha'], data['cedula'], data['nombre'], data['turno'], data['h_inicio'], data['h_fin'],
+                 data['concesion'], data['control'], ruta.strip(), data['linea'], data['cop'], data['observaciones'],
+                 data['usuario_registra'], data['registrado_por'], data['fecha_hora_registro'], data['puestosSC'], 
+                 data['puestosUQ'], data['cedula_enlace'], data['nombre_supervisor_enlace'])
+                for data in processed_data for ruta in data['ruta'].split(',')
+            ]
 
-                    cursor.execute('''
-                        INSERT INTO asignaciones (fecha, cedula, nombre, turno, h_inicio, h_fin, concesion, control, ruta, linea, cop, observaciones, usuario_registra, registrado_por, fecha_hora_registro, puestosSC, puestosUQ, cedula_enlace, nombre_supervisor_enlace)
-                        VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?)
-                    ''', (
-                        data['fecha'], data['cedula'], data['nombre'], data['turno'], data['h_inicio'], data['h_fin'],
-                        data['concesion'], data['control'], ruta, data['linea'], data['cop'], data['observaciones'],
-                        data['usuario_registra'], data['registrado_por'], data['fecha_hora_registro'], data['puestosSC'], data['puestosUQ'],
-                        data['cedula_enlace'], data['nombre_supervisor_enlace']
-                    ))
-                    print(f"Nuevo registro insertado para cédula: {data['cedula']}, fecha: {data['fecha']}, ruta: {ruta}")
+            cursor.executemany('''
+                INSERT INTO asignaciones (fecha, cedula, nombre, turno, h_inicio, h_fin, concesion, control, ruta, linea, cop, observaciones, usuario_registra, registrado_por, fecha_hora_registro, puestosSC, puestosUQ, cedula_enlace, nombre_supervisor_enlace)
+                VALUES (%s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s)
+            ''', insert_values)
 
             # Confirmar la transacción
             conn.commit()
             return {"status": "success", "message": "Asignaciones guardadas exitosamente."}
 
-        except sqlite3.Error as e:
-            # Revertir la transacción en caso de error
+        except psycopg2.Error as e:
             conn.rollback()
             print(f"Error al guardar asignaciones: {e}")
             raise HTTPException(status_code=500, detail=f"Error al guardar asignaciones: {e}")
